# openapi.yaml
openapi: 3.0.0
info:
  title: After WOD API
  description: CrossFit 보조 운동 계획 생성 및 운동 기록 관리 API
  version: 1.0.0
  contact:
    name: Hound ANAS
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: 개발 서버
...
paths:
...
  /api/v1/workouts/generate/wod:
    post:
      tags:
        - 운동 계획
      summary: 운동 계획 생성 - WOD 모드 (부족한 부위 채우기)
      description: |
        오늘 수행한 WOD를 기반으로 부족한 부위를 채울 운동 계획을 생성합니다.
        생성된 계획은 데이터베이스에 저장되지 않으며, 사용자가 운동을 시작할 때 저장됩니다.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutGenerateWodRequest'
            example:
              duration: 20
              wodMovementIds: [mov1, mov5, mov10]
      responses:
        '200':
          description: 운동 계획 생성 성공 (임시 데이터)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutGenerateResponse'
        '400':
          description: 요청 데이터 유효성 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequestResponse'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorizedResponse'
  /api/v1/workouts/generate/goal:
    post:
      tags:
        - 운동 계획
      summary: 운동 계획 생성 - GOAL 모드 (나의 달성 목표)
      description: |
        달성하고 싶은 스킬을 기반으로 목표 달성을 위한 운동 계획을 생성합니다.
        사용자의 난이도에 맞는 템플릿을 사용하여 계획을 생성합니다.
        생성된 계획은 데이터베이스에 저장되지 않으며, 사용자가 운동을 시작할 때 저장됩니다.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutGenerateGoalRequest'
            example:
              duration: 20
              goalMovementId: muscle_up
              difficultyLevel: BEGINNER
      responses:
        '200':
          description: 운동 계획 생성 성공 (임시 데이터)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutGenerateResponse'
        '400':
          description: 요청 데이터 유효성 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequestResponse'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorizedResponse'
  /api/v1/workouts/generate/part:
    post:
      tags:
        - 운동 계획
      summary: 운동 계획 생성 - PART 모드 (타겟 부위 선택)
      description: |
        선택한 부위를 집중적으로 운동할 계획을 생성합니다.
        생성된 계획은 데이터베이스에 저장되지 않으며, 사용자가 운동을 시작할 때 저장됩니다.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutGeneratePartRequest'
            example:
              duration: 20
              targetMuscleGroups: [CORE, CHEST]
      responses:
        '200':
          description: 운동 계획 생성 성공 (임시 데이터)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutGenerateResponse'
        '400':
          description: 요청 데이터 유효성 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequestResponse'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorizedResponse'
...
components:
  schemas:
    CommonResult:
      type: object
      description: 모든 API 응답의 공통 포맷
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          description: 요청 성공 여부
        message:
          type: string
          description: 응답 메시지
        data:
          type: object
          description: 실제 응답 데이터 (실패 시 에러 정보, 성공 시 비즈니스 데이터)
          nullable: true

    # 응답 스키마들
...
    ErrorUnauthorizedResponse:
      allOf:
        - $ref: '#/components/schemas/CommonResult'
      example:
        success: false
        message: 인증 실패
        data: null
    ErrorBadRequestResponse:
      allOf:
        - $ref: '#/components/schemas/CommonResult'
      example:
        success: false
        message: 요청 데이터가 유효하지 않습니다.
        data: null
...
    WorkoutGenerateResponse:
      allOf:
        - $ref: '#/components/schemas/CommonResult'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/WorkoutPlan'
      example:
        success: true
        message: 운동 계획이 생성되었습니다.
        data:
          mode: WOD
          duration: 20
          rounds: 6
          targetTimePerRound: 3분 20초
          exercises:
            - exerciseId: 0
              movementId: mov1
              name: 할로우 락
              minReps: 10
              maxReps: 15
              equipment:
                - BODYWEIGHT
              muscleGroups:
                - CORE
            - exerciseId: 1
              movementId: mov2
              name: 플랭크
              minReps: 30
              maxReps: 60
              equipment:
                - BODYWEIGHT
              muscleGroups:
                - CORE
          modeDisplay: 부족 부위 채우기
...
    WorkoutGenerateWodRequest:
      type: object
      required:
        - duration
      properties:
        duration:
          type: integer
          description: 운동 시간 (분)
          enum:
            - 1
            - 5
            - 10
            - 15
            - 20
            - 25
            - 30
        wodMovementIds:
          type: array
          items:
            type: string
          description: 오늘 수행한 운동 ID 목록
    WorkoutGenerateGoalRequest:
      type: object
      required:
        - duration
        - goalMovementId
        - difficultyLevel
      properties:
        duration:
          type: integer
          description: 운동 시간 (분)
          enum:
            - 1
            - 5
            - 10
            - 15
            - 20
            - 25
            - 30
        goalMovementId:
          type: string
          description: 달성하고 싶은 스킬의 운동 ID
        difficultyLevel:
          type: string
          description: 난이도 (초보자/중급자/상급자)
          enum:
            - BEGINNER
            - INTERMEDIATE
            - ADVANCED
    WorkoutGeneratePartRequest:
      type: object
      required:
        - duration
        - targetMuscleGroups
      properties:
        duration:
          type: integer
          description: 운동 시간 (분)
          enum:
            - 1
            - 5
            - 10
            - 15
            - 20
            - 25
            - 30
        targetMuscleGroups:
          type: array
          description: 타겟 근육 그룹
          items:
            type: string
            enum:
              - CORE
              - LEGS
              - BACK
              - CHEST
              - SHOULDER
              - ARMS
              - CARDIO
              - FULL
...
    WorkoutPlan:
      type: object
      description: 생성된 운동 계획 (임시 데이터, DB에 저장되지 않음)
      required:
        - mode
        - duration
        - rounds
        - exercises
      properties:
        mode:
          type: string
          description: 운동 모드
          enum:
            - WOD
            - GOAL
            - PART
        duration:
          type: integer
          description: 운동 시간 (분)
        rounds:
          type: integer
          description: 라운드 수
        targetTimePerRound:
          type: string
          description: 라운드당 목표 시간
        exercises:
          type: array
          items:
            $ref: '#/components/schemas/Exercise'
          description: 운동 목록
        modeDisplay:
          type: string
          description: 모드 표시 텍스트
    Exercise:
      type: object
      required:
        - exerciseId
        - movementId
        - name
      properties:
        exerciseId:
          type: integer
          description: 이 계획 내에서의 운동 순서 (0부터 시작)
        movementId:
          type: string
          description: 운동 ID (movements 테이블 참조)
        name:
          type: string
          description: 운동 이름
        minReps:
          type: integer
          description: 최소 반복 횟수
        maxReps:
          type: integer
          description: 최대 반복 횟수
        equipment:
          type: array
          description: 필요 장비
          items:
            type: string
            enum:
              - BODYWEIGHT
              - GHD
              - BAR
              - BAND
              - RINGS
              - BARBELL
              - BOX
              - DUMBBELL
              - KETTLEBELL
              - WALLBALL
              - WALL
              - ASSAULT_BIKE
              - ROWING
        muscleGroups:
          type: array
          description: 타겟 근육 그룹
          items:
            type: string
            enum:
              - CORE
              - LEGS
              - BACK
              - CHEST
              - SHOULDER
              - ARMS
              - CARDIO
              - FULL
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 기반 인증 (로그인 후 발급받은 accessToken 사용)
security:
  - BearerAuth: []